+++ examples/1c4a-example.js
@@ -1,6 +1,4 @@
-const CountryCode = require('../lib/1c4a/countryCode');
-const { VoucherLayout } = require('../lib/1c4a/voucher');
-const { Internetmarke } = require('../lib/Internetmarke');
+const { CountryCode, Internetmarke, VoucherLayout } = require('..');
 
 // To start this example directly make sure to build the project first:
 // npm i && npm run build