+++ index.js
@@ -7,3 +7,4 @@
 'use strict';
 
 module.exports = require('./lib/Internetmarke');
+